Back-End/requirements.txt
fastapi==0.119.0
uvicorn==0.37.0
pydantic>=2.0
pydantic-settings>=2.0
python-dotenv==1.1.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
boto3>=1.34.0
mangum>=0.17.0
openai>=1.0.0
